requirements.txt
flask==3.0.0
python-pptx==0.6.23
werkzeug==3.0.1